Correctly redirect from all Actions
What happened is that after every server action,
remix redirected to "/" route.
But because the "/" route does not exists,
user was redirected again to /resume so
it seemed that everything is fine.

Now every server action just redirects back to
where the user was before. This has benefits:

* there is only one redirect
(previously firstly to "/", then to "/resume")

* it's possible to add new pages and actions will
still work properly (instead of always redirecting to /resume)